requirements.txt
# Core dependencies
python>=3.9

# Deep Learning & ML
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
xgboost>=2.0.0
scikit-learn>=1.3.0

# LLM APIs
google-generativeai>=0.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0
pyarrow>=12.0.0

# Knowledge Tracing
pykt-toolkit>=0.0.30

# Training utilities
accelerate>=0.24.0
peft>=0.6.0
trl>=0.7.4
bitsandbytes>=0.41.0

# Config & logging
pyyaml>=6.0
python-dotenv>=1.0.0
wandb>=0.15.0
tqdm>=4.66.0

# Utilities
requests>=2.31.0
beautifulsoup4>=4.12.0